+++ ocaml-lsp-server/src/config_data.ml
@@ -78,6 +78,7 @@ module InlayHints = struct
     { hint_pattern_variables : bool [@key "hintPatternVariables"] [@default false]
     ; hint_let_bindings : bool [@key "hintLetBindings"] [@default false]
     ; hint_function_params : bool [@key "hintFunctionParams"] [@default true]
+    ; hint_avoid_ghost_locations : bool [@key "hintAvoidGhostLocations"] [@default true]
     }
   [@@deriving_inline yojson] [@@yojson.allow_extra_fields]
 
@@ -90,6 +91,7 @@ module InlayHints = struct
        let hint_pattern_variables_field = ref Ppx_yojson_conv_lib.Option.None
        and hint_let_bindings_field = ref Ppx_yojson_conv_lib.Option.None
        and hint_function_params_field = ref Ppx_yojson_conv_lib.Option.None
+       and avoid_ghost_locations_field = ref Ppx_yojson_conv_lib.Option.None
        and duplicates = ref []
        and extra = ref [] in
        let rec iter = function
@@ -116,6 +118,13 @@ module InlayHints = struct
                  hint_function_params_field := Ppx_yojson_conv_lib.Option.Some fvalue
                | Ppx_yojson_conv_lib.Option.Some _ ->
                  duplicates := field_name :: Ppx_yojson_conv_lib.( ! ) duplicates)
+            | "hintAvoidGhostLocations" ->
+              (match Ppx_yojson_conv_lib.( ! ) avoid_ghost_locations_field with
+               | Ppx_yojson_conv_lib.Option.None ->
+                 let fvalue = bool_of_yojson _field_yojson in
+                 avoid_ghost_locations_field := Ppx_yojson_conv_lib.Option.Some fvalue
+               | Ppx_yojson_conv_lib.Option.Some _ ->
+                 duplicates := field_name :: Ppx_yojson_conv_lib.( ! ) duplicates)
             | _ -> ());
            iter tail
          | [] -> ()
@@ -137,11 +146,13 @@ module InlayHints = struct
            | [] ->
              let ( hint_pattern_variables_value
                  , hint_let_bindings_value
-                 , hint_function_params_value )
+                 , hint_function_params_value
+                 , avoid_ghost_locations_value )
                =
                ( Ppx_yojson_conv_lib.( ! ) hint_pattern_variables_field
                , Ppx_yojson_conv_lib.( ! ) hint_let_bindings_field
-               , Ppx_yojson_conv_lib.( ! ) hint_function_params_field )
+               , Ppx_yojson_conv_lib.( ! ) hint_function_params_field
+               , Ppx_yojson_conv_lib.( ! ) avoid_ghost_locations_field )
              in
              { hint_pattern_variables =
                  (match hint_pattern_variables_value with
@@ -153,6 +164,10 @@ module InlayHints = struct
                   | Ppx_yojson_conv_lib.Option.Some v -> v)
              ; hint_function_params =
                  (match hint_function_params_value with
+                  | Ppx_yojson_conv_lib.Option.None -> true
+                  | Ppx_yojson_conv_lib.Option.Some v -> v)
+             ; hint_avoid_ghost_locations =
+                 (match avoid_ghost_locations_value with
                   | Ppx_yojson_conv_lib.Option.None -> true
                   | Ppx_yojson_conv_lib.Option.Some v -> v)
              }))
@@ -168,6 +183,7 @@ module InlayHints = struct
      | { hint_pattern_variables = v_hint_pattern_variables
        ; hint_let_bindings = v_hint_let_bindings
        ; hint_function_params = v_hint_function_params
+       ; hint_avoid_ghost_locations = v_hint_avoid_ghost_locations
        } ->
        let bnds : (string * Ppx_yojson_conv_lib.Yojson.Safe.t) list = [] in
        let bnds =
@@ -181,6 +197,10 @@ module InlayHints = struct
        let bnds =
          let arg = yojson_of_bool v_hint_pattern_variables in
          ("hintPatternVariables", arg) :: bnds
+       in
+       let bnds =
+         let arg = yojson_of_bool v_hint_avoid_ghost_locations in
+         ("hintAvoidGhostLocations", arg) :: bnds
        in
        `Assoc bnds
      : t -> Ppx_yojson_conv_lib.Yojson.Safe.t)
@@ -929,6 +949,7 @@ let default =
         { hint_pattern_variables = false
         ; hint_let_bindings = false
         ; hint_function_params = false
+        ; hint_avoid_ghost_locations = true
         }
   ; dune_diagnostics = Some { enable = true }
   ; syntax_documentation = Some { enable = false }